pull validator into separate file + move team checks to validator
